requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
geopy
Pillow
